Move everything from gaia.core to gaia.__init__, but keep some imports in gaia.core for backward compatibility with existing plugins
