+++ requirements.txt
@@ -19,7 +19,7 @@ argparse==1.3.0
 shapely==1.5.13
 geopandas==0.1.1
 pillow==3.1.1
-gdal==1.10
+gdal==1.11
 celery==3.1.20
 psycopg2==2.6.1
 geoalchemy2==0.2.6
